Hmm.. the queries generated through a single query obviously suck balls, and couldn't find a better way to do it, than breaking it down again into a first query to fethc IDs only with paged results (therefore the pager object needs to be passed manually to templates), and a second query to do the actual hit-counting via group-by and fetching the extra records when needed (eg. user info).
Imho, this is still not good.. if the URLs in the set had a very big
number of hits, performance would suck hairy donkey balls again, but the
way to go if I were really aiming to solve such scale problems would
probably be much different, and keeping the hit-count and some other
views materialized on other tables (eg. with breakdowns on the different
fields, rollups, etc.). not yet the case.